feat: Add skip logic to ignore datasets marked for skipping
This commit adds skip logic to ignore datasets that are marked for skipping. This logic is implemented in multiple files: `0_inspect.py`, `1_dataset2bids.py`, `3b_prepInspection.py`, `4_features.py`, `5_aggregate.py`, and `2_participants.py`. The logic checks if the dataset has a `skip` flag set to `True` and continues to the next dataset if it does. This allows for more efficient processing by skipping unnecessary datasets.

Recent user commits:
- Refactor parallelization logic for feature extraction
- skip read eeg on done files
- feature parallelization
- ROI features
- preprocessing run
- viz v1
- combat notebook notes
- Add files via upload (full code of features)
- bug fixing
- roi aggregate support, bug fix, norm after prep